requirements.txt
grpcio
grpcio-tools
opencv-python
pillow
numpy
torch
pandas
torchvision
